buybutton

+++ Assets/Scripts/Stock/StockList/Stock_Samsung.cs
@@ -6,7 +6,7 @@ public class Stock_Samsung : MonoBehaviour
 {
     public static int stock_id = 0;
     public static string stockName = "»ï¼º";
-    public static float price = 50000f;
+    public static int price = 50000;
     public static List<int> stockValue = new List<int> { 10, 100, 1000, 10000, 100000 };
     protected float changeRange;
     protected bool isIncrease;
